refactor(agent-ui): ChatPanel partagé — AgentPage et EntityAssistant dédupliqués
~150 lignes dupliquées entre les deux surfaces de chat (types Message,
transcript optimiste, submit + retry, undo créations/modifications,
privacy gate, form) : toute amélioration devait être faite deux fois.

- Nouveau ChatPanel : transcript + composer + toute la logique partagée,
  paramétré par testIdPrefix / empty state / className
- create-on-first-message injecté via ensureConversation (AgentPage) ;
  absent = panel désactivé tant que la conversation ancrée charge
  (EntityAssistant)
- AgentPage ne garde que la sidebar + la sélection de conversation ;
  EntityAssistant devient ~30 lignes

Aucun changement de comportement — testids identiques, E2E agent
10/10, prérequis du streaming (palier 3).
